fix：UI解绑定，若上次运行报错未解绑则报错则清除

+++ Assets/Scripts/NM/Data/FSM/Playing/GamePlaying.cs
@@ -11,7 +11,7 @@ public partial class GamePlaying : GameRoot.StateFSM<GamePlaying>
 {
     public override string ToString() => nameof(GamePlaying);
     public string PlayerName = "Deli";
-    public float PlayTime;
+    public double PlayTime;
     public List<SymbolData> SymbolDeckList = [];
     public long Coin;
     public int RemoveToken;
